Move memory to top-level package
Now that we have direct access to addr and malloc,
it's only logical to also expose memory as an (unsafe)
part of api.
