+++ core/src/main/scala/offheap/Memory.java
@@ -1,26 +1,9 @@
 package offheap.internal;
 
-import sun.misc.Unsafe;
-import java.lang.reflect.Field;
 import static offheap.internal.Sanitizer.validate;
+import static offheap.internal.SunMisc.UNSAFE;
 
 public class Memory {
-    public static final Unsafe UNSAFE;
-    static {
-        Unsafe value = null;
-        try {
-            value = Unsafe.getUnsafe();
-        } catch (SecurityException e) {
-            try {
-                Field f = Unsafe.class.getDeclaredField("theUnsafe");
-                f.setAccessible(true);
-                value = (Unsafe) f.get(null);
-            } catch (NoSuchFieldException nsfe) {
-            } catch (IllegalAccessException iae) {
-            }
-        }
-        UNSAFE = value;
-    }
     public static long allocate(long size) {
         return UNSAFE.allocateMemory(size);
     }
