+++ core/src/main/scala/offheap/malloc.scala
@@ -1,6 +1,6 @@
 package offheap
 
-import offheap.internal.Memory.UNSAFE
+import offheap.internal.SunMisc.UNSAFE
 
 /** Underyling OS allocator that does not attempt
  *  to perform any automatic memory management
